+++ src/main/java/com/hollroom/monthly/domain/entity/MonthlyProductEntity.java
@@ -22,7 +22,7 @@ public class MonthlyProductEntity {
 //    private UserEntity user;
     private int deposit;
     private int monthly;
-    private int divisionCode;
+    private Long divisionCode;
     private int floorCount;
     private int pyeongCount;
     private int roomCount;
@@ -36,7 +36,7 @@ public class MonthlyProductEntity {
     private Date updatedAt;
     private Date deletedAt;
 
-    public MonthlyProductEntity(int deposit, int monthly, int divisionCode, int floorCount, int pyeongCount, int roomCount, int bayCount, String roomOption, String securityFacility, Date expirationDate) {
+    public MonthlyProductEntity(int deposit, int monthly, Long divisionCode, int floorCount, int pyeongCount, int roomCount, int bayCount, String roomOption, String securityFacility, Date expirationDate) {
         this.deposit = deposit;
         this.monthly = monthly;
         this.divisionCode = divisionCode;
